Task 65004: Alterar endpoint Download para enviar para fila no worker

+++ src/SME.SERAp.Prova.Api/Controllers/DownloadController.cs
@@ -25,7 +25,7 @@ namespace SME.SERAp.Prova.Api.Controllers
         [ProducesResponseType(typeof(bool), 200)]
         [ProducesResponseType(typeof(RetornoBaseDto), 500)]
         [ChaveAutenticacaoApi]
-        public async Task<IActionResult> ExcluirDownloads(int[] ids, [FromServices] IExcluirDownloadProvaAlunoUseCase excluirDownloadsProvaAlunoUseCase)
+        public async Task<IActionResult> ExcluirDownloads(long[] ids, [FromServices] IExcluirDownloadProvaAlunoUseCase excluirDownloadsProvaAlunoUseCase)
         {
             return Ok(await excluirDownloadsProvaAlunoUseCase.Executar(ids));
         }
